pyproject.toml
[project]
name = "schema-bridge"
version = "0.1.0"
description = "Standalone RDF canonicalization and transformation pipeline for EMX2."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "rdflib>=7.0.0",
  "gql[requests]>=3.5.0",
  "pyyaml>=6.0.0",
  "pyshacl>=0.25.0",
  "morph-kgc>=2.8.0",
  "typer>=0.12.0",
]

[project.optional-dependencies]
test = [
  "pytest>=8.0.0",
]

[project.scripts]
schema-bridge = "schema_bridge.cli:main"
schema-bridge-ingest = "schema_bridge.ingest_cli:main"

[build-system]
requires = ["hatchling>=1.24.0"]
build-backend = "hatchling.build"

[tool.hatch.build]
include = [
  "src/schema_bridge/**",
]

[tool.hatch.build.targets.wheel]
packages = ["src/schema_bridge"]
